feat: 살까말까 api 기능 구현 ing
id long 으로 변경, readme 작성, logback 설정, 패키지 정리

#8

+++ src/main/java/com/stocking/infra/common/BaseEntity.java
@@ -18,14 +18,14 @@ import lombok.Data;
 public abstract class BaseEntity {
 
     @Column(name = "created_id")
-    private int createdId;
+    private long createdId;
 
     @CreatedDate
     @Column(name = "created_date")
     private LocalDateTime createdDate;
 
     @Column(name = "updated_id")
-    private int updatedId;
+    private long updatedId;
 
     @LastModifiedDate
     @Column(name = "updated_date")
